% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{ask}
\alias{ask}
\title{Process a Prompt using Specified API}
\usage{
ask(prompt, system = NULL, model = "claude", local = FALSE, ...)
}
\arguments{
\item{prompt}{A character string containing the user's message.}

\item{system}{An optional character string containing the system message. Defaults to NULL.}

\item{model}{A character string specifying the model or API to use.}

\item{local}{Logical indicating whether to use a local LLM via Ollama server or not. Default is FALSE.}

\item{...}{Additional arguments passed to the specific API functions.}
}
\value{
A character string containing the response from the chosen API.
}
\description{
Process a Prompt using Specified API
}
