% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{ask_google}
\alias{ask_google}
\title{Call the Google Gemini API to Process a Prompt}
\usage{
ask_google(
  prompt,
  system = NULL,
  model = "gemini-1.5-pro",
  temperature = 0.7,
  ...
)
}
\arguments{
\item{prompt}{A character string containing the user's message to be processed by the Google API.}

\item{system}{An optional character string to provide a system prompt to the model. Default is NULL.}

\item{model}{A character string specifying the model to use. Default is "gemini-1.5-pro".}

\item{temperature}{A numeric value representing the temperature parameter for the API call. Default is 0.7.}

\item{...}{Additional parameters to pass to the Google API.}
}
\value{
A character string containing the processed text from the Google API response. If an error occurs, the function will return NULL.
}
\description{
This function uses the Google Gemini API to process a given prompt.
}
