% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{ask_groq}
\alias{ask_groq}
\title{Call the Groq API to Process a Prompt}
\usage{
ask_groq(
  prompt,
  system = NULL,
  model = "groq",
  temperature = 0.3,
  max_tokens = 4096,
  ...
)
}
\arguments{
\item{prompt}{A character string containing the user's message to be processed by the Groq API.}

\item{system}{An optional character string to provide a system prompt to the model. Default is NULL.}

\item{model}{A character string specifying the model to use. Can be generic ("groq", "mixtral", "llama") or specific (e.g., "mixtral-8x7b-32768").}

\item{temperature}{A numeric value representing the temperature parameter for the API call. Default is 0.3.}

\item{max_tokens}{An integer specifying the maximum number of tokens in the response. Default is 4096.}

\item{...}{Additional parameters to pass to the Groq API.}
}
\value{
A character string containing the processed text from the Groq API response. If an error occurs, the function will return NULL.
}
\description{
This function uses the Groq API to process a given prompt. It supports various open-source models
available through Groq's API.
}
