% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_pdf_block}
\alias{create_pdf_block}
\title{Create PDF content block for API request}
\usage{
create_pdf_block(pdf_data, cache = FALSE)
}
\arguments{
\item{pdf_data}{Base64 encoded PDF data}

\item{cache}{Logical indicating whether to cache the PDF. Default is FALSE}
}
\value{
List containing the PDF content block
}
\description{
Create PDF content block for API request
}
