% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{encode_pdf}
\alias{encode_pdf}
\title{Read and encode PDF file}
\usage{
encode_pdf(path)
}
\arguments{
\item{path}{Path to the PDF file or URL}
}
\value{
Base64 encoded string of the PDF content
}
\description{
Read and encode PDF file
}
