% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{imagine_openai}
\alias{imagine_openai}
\title{Process a Prompt using OpenAI's DallE Models}
\usage{
imagine_openai(
  prompt,
  model = "dall-e-3",
  image_size = "1792x1024",
  image_style = "vivid",
  output = "url"
)
}
\arguments{
\item{prompt}{A character string containing the user's message to the GPT model.}

\item{model}{A character string containing the GPT model to use.}

\item{image_size}{Default is "1792x1024" (but square is faster).}

\item{image_style}{A character string containing the style to use (vivid or natural).}

\item{output}{Default is url.}
}
\value{
A character string containing the GPT model's response.
}
\description{
This function uses the OpenAI API to interactively communicate with a specified DallE model.
The default model is DallE3, but it can use DallE2.
}
