% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{ask}
\alias{ask}
\title{Process a Prompt using Specified API}
\usage{
ask(prompt, system = NULL, model = "gpt-3.5-turbo", ...)
}
\arguments{
\item{prompt}{A character string containing the user's message.}

\item{system}{An optional character string containing the system message. Defaults to NULL.}

\item{model}{A character string specifying the model or API to use.
Can be a general API name (e.g., "anthropic", "google", "openai", "groq")
or a specific model version (e.g., "gpt-4", "claude-3-opus-20240229").}

\item{...}{Additional arguments passed to the specific API functions.}
}
\value{
A character string containing the response from the chosen API.
}
\description{
This function acts as a wrapper to process a prompt using Anthropic, Google, Groq, or OpenAI's API based on the specified model.
}
