% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{ask_anthropic}
\alias{ask_anthropic}
\title{Call Anthropic API}
\usage{
ask_anthropic(
  prompt,
  system = NULL,
  model = "claude",
  temperature = 0,
  max_tokens = 4096,
  pre_fill = NULL,
  ...
)
}
\arguments{
\item{prompt}{A character string representing the prompt to be sent to the Claude API.}

\item{system}{An optional character string to provide a system prompt to the model. Default is NULL.}

\item{model}{A character string specifying the model to use. Can be generic ("claude", "anthropic") or specific (e.g., "claude-3-opus-20240229").}

\item{temperature}{A numeric value representing the temperature parameter for the API call. Default is 0.}

\item{max_tokens}{An integer specifying the maximum number of tokens in the response. Default is 4096.}

\item{pre_fill}{An optional character string to pre-fill the model's response. Default is NULL.}

\item{...}{Additional parameters to pass to the Anthropic API.}
}
\value{
The generated text response from the Claude API, or NULL if an error occurs.
}
\description{
This function sends a prompt to the Claude API by Anthropic and returns the generated text response.
}
